pi_django/requirements.txt
Django==2.2.1
cryptography==2.6.1
djangorestframework==3.9.3
